requirements.txt
fastapi
uvicorn
line-bot-sdk
google-generativeai>=0.7.2